+++ components/pm-subscription-handler/Changelog.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 * Updated PMSH app configuration, simplified existing config (DCAEGEN2-2814)
 * Created Schema definitions in swagger file according to the new structure (DCAEGEN2-2889)
 * Implemented Create Subscription public API (DCAEGEN2-2819)
+* Added 2 new attributes to the subscription model (DCAEGEN2-2913)
 
 ## [1.3.2]
 ### Changed